requirements.txt
aioredis==1.3.1
asgiref==3.3.4
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
Django==3.2
django-appconf==1.0.4
django-fsm==2.7.1
django-fsm-log==2.0.1
django-sockpuppet==0.5.2
graphviz==0.16
hiredis==2.0.0
hyperlink==21.0.0
idna==3.1
incremental==21.3.0
msgpack==1.0.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
pytz==2021.1
service-identity==18.1.0
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
Twisted==21.2.0
txaio==21.2.1
zope.interface==5.3.0
